I want to use this file to add major changes or experimenting before pushing into a clean final code.
For instance there is multiple producers of energy and it seems some may be focused on green energy more than others. If true, we can use this as reason for focused promotion towards a green energy audience.

This is one example we can have to set the bar on this final project.